major update

+++ src/services/rental.ts
@@ -1,5 +1,6 @@
 import RENTAL_DATA from "@/interfaces/rental";
 import Rental from "@/models/rental";
+import { Op, WhereOptions } from "sequelize";
 
 export const findOneRentalRecord = (
   input: Partial<RENTAL_DATA>
@@ -12,10 +13,23 @@ export const findOneRentalRecord = (
   });
 };
 
-export const showAllRental = (uuid?: string): Promise<Rental[]> => {
+type RENTAL_OPTS = {
+  offset: number;
+};
+
+export const findAllRentalRecord = (
+  input: Partial<Rental>,
+  opts: RENTAL_OPTS = { offset: 0 }
+): Promise<Rental[]> => {
+  let inputName: WhereOptions<any> | undefined;
+  if (input.name)
+    inputName = { ...input, name: { [Op.like]: "%" + input.name + "%" } };
   return new Promise((resolve, reject) => {
-    const where = uuid ? { id_owner: uuid } : undefined;
-    Rental.findAll({ where })
+    Rental.findAll({
+      where: inputName ? inputName : input,
+      offset: opts.offset,
+      limit: 20,
+    })
       .then((result) => resolve(result))
       .catch((err) => reject(err));
   });
@@ -29,7 +43,9 @@ export const createRental = (input: RENTAL_DATA): Promise<Rental> => {
   });
 };
 
-export const editRental = (input: RENTAL_DATA): Promise<number | undefined> => {
+export const updateRentalRecord = (
+  input: RENTAL_DATA
+): Promise<number | undefined> => {
   return new Promise((resolve, reject) => {
     Rental.update(input, { where: { id: input.id, id_owner: input.id_owner } })
       .then((result) => resolve(result[0]))
@@ -37,9 +53,12 @@ export const editRental = (input: RENTAL_DATA): Promise<number | undefined> => {
   });
 };
 
-export const deleteRental = (id: string): Promise<number | undefined> => {
+export const deleteOneRentalRecord = (input: {
+  id: number;
+  id_owner: string;
+}): Promise<number | undefined> => {
   return new Promise((resolve, reject) => {
-    Rental.destroy({ where: { id } })
+    Rental.destroy({ where: { id: input.id, id_owner: input.id_owner } })
       .then((result) => resolve(result))
       .catch((err) => reject(err));
   });
